requirements.txt
accelerate==0.34.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.114.2
ffmpy==0.4.0
filelock==3.13.1
flash-attn==2.6.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
gradio==3.23.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.4
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
latex2mathml==3.77.0
linkify-it-py==2.0.3
lit==15.0.7
-e git+https://github.com/microsoft/LLaVA-Med.git@b9a98a736d2ef05bcf5ff345be6403fb3a664eaf#egg=llava
markdown-it-py==2.2.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.8.1
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.3
open_clip_torch==2.26.1
openai==0.27.8
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.2.0
platformdirs==4.3.3
protobuf==5.28.1
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.28.1
rpds-py==0.20.0
safetensors==0.4.5
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.5
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
timm==1.0.9
tokenizers==0.12.1
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tqdm==4.66.5
transformers @ git+https://github.com/huggingface/transformers@cae78c46d658a8e496a815c2ee49b9b178fb9c9a
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==1.26.13
uvicorn==0.30.6
wandb==0.18.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==13.0.1
yarl==1.11.1
